refactor: App.jsx내부 로직을 useAppLogic 훅으로 분리하여 코드 가독성 및 재사용성 향상
